+++ CHANGELOG.md
@@ -13,7 +13,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ### Compatible changes
 
--
+- Treat transient factory_bot attributes as associations if they are named like a factory.
 
 ## 2.4.1 - 2022-03-16